requirements.txt
flask
whisper
openai-whisper
gunicorn
ffmpeg-python
python-dotenv